cleaning up FolderRegionProvider's Iterators
